% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_caption.R
\name{add_caption}
\alias{add_caption}
\title{Add caption as custom style}
\usage{
add_caption(style = "Image Caption", caption = "caption")
}
\arguments{
\item{style}{custom style name}

\item{caption}{caption text}
}
\description{
To adding caption with custom style.
}
\details{
Adding caption with custom style. Custom style is produced by Pandoc. Use inline R code chunk instead of R code block.
}
