% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fieldcode.R
\name{add_fieldcode}
\alias{add_fieldcode}
\title{Add any field code}
\usage{
add_fieldcode(code, text)
}
\arguments{
\item{code}{filed code}

\item{text}{default text}
}
\description{
General use of adding field code
}
\details{
hoge
}
