% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_figure.R
\name{add_figure}
\alias{add_figure}
\title{Add figure as a Markdown code}
\usage{
add_figure(caption, file, width = NULL)
}
\arguments{
\item{caption}{caption for image}

\item{file}{file path (absolute path)}

\item{width}{figure's width (optional)}
}
\description{
Adding markdown code of figure/image.
}
\details{
To add figures with any size easily, use add_figure() with inline R chunk code. This is for escaping that knitr::include_graphics() generate comment after markdown. Because of this unneeded comment, captioner cannot be worked well. This function will produce simple markdown notation of image/figure. This can be used for captioner.
}
