% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpage.R
\name{newpage}
\alias{newpage}
\title{Insert New Page in docx}
\usage{
newpage()
}
\description{
Insert new page by adding ooxml code into docx. This function returns the ooxml codes without any parameters. Use inline R chunk insetead of R code block (to add ooxml code directly into sentenses)
}
