% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_space_preserve.R
\name{text_space_preserve}
\alias{text_space_preserve}
\title{Text with white space}
\usage{
text_space_preserve(text, lang = "ja-JP")
}
\arguments{
\item{text}{text including white space}

\item{lang}{language. this is not valid param.}
}
\description{
Insert text with preserved spaces.
}
\details{
In Pandoc's transformation, white space at a beginning of sentense will be ignored. By using this function, you can get the white space preserved sentense.
}
